# [BEAM-12640] Enable adding JHM to Java based projects and add an example benchmark to the SDK harness.