--- a/sdks/java/build-tools/src/main/resources/beam/spotbugs-filter.xml
+++ b/sdks/java/build-tools/src/main/resources/beam/spotbugs-filter.xml
@@ -64,6 +64,9 @@
     <Class name="~.*AutoValue_.*"/>
   </Match>
   <Match>
+    <Package name="~.*jmh_generated.*"/>
+  </Match>
+  <Match>
     <Package name="org.apache.beam.sdk.extensions.sql.impl.parser.impl"/>
   </Match>
   <Match>