@lukecwik 2021-07-21
Run Java PreCommit